refactor(windows): share git-path normalization helper
